refactor local server code to src/server
Adding a bit of organization to make it easier
to add more pages and apis
