+++ cli.js
@@ -298,7 +298,7 @@ const cli = {
   },
   // run the dev server
   server: () => {
-    require('./server.js');
+    require('./src/server/server.js');
   },
   // help section
   help: (command) => {
